Updated to match comments
