+++ server/src/config.ts
@@ -1,22 +1,7 @@
 import { Request } from "express";
 import type { Knex } from "knex";
-import { createLogger, Logger, transports, format } from "winston";
-const { combine, timestamp, splat, printf, label } = format;
 
-const level = process.env.LOG_LEVEL || "info";
-
-const timestampPrefixLogFormat = printf(
-  ({ level, message, label, timestamp }) => {
-    return `${timestamp} [${label}] ${level}: ${message}`;
-  }
-);
-
-const customFormatWithTimestamp = combine(
-  label({ label: "univaf" }),
-  timestamp(),
-  splat(),
-  timestampPrefixLogFormat
-);
+export const LOG_LEVEL = process.env.LOG_LEVEL || "info";
 
 export function getApiKeys(): Array<string> {
   let keyList = process.env.API_KEYS;
@@ -51,18 +36,4 @@ export function loadDbConfig(): Knex.Config {
   const knexfile = require("../knexfile");
   const nodeEnv = process.env.NODE_ENV || "development";
   return knexfile[nodeEnv];
-}
-
-export const logger: Logger = createLogger({
-  format: customFormatWithTimestamp,
-  level: level,
-  transports: [new transports.Console()],
-});
-
-export function logStackTrace(logger: Logger, err: any): void {
-  if (err instanceof Error) {
-    logger.error(`${err.stack || err}`);
-  } else {
-    logger.error(err);
-  }
 }
